Move shim code around to avoid package import cycles
