+++ pf/tfgen/schemashim.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2022, Pulumi Corporation.
+// Copyright 2016-2023, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,18 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package schemashim
+package tfgen
 
 import (
 	"context"
 
-	pfprovider "github.com/hashicorp/terraform-plugin-framework/provider"
-	shim "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfshim"
+	"github.com/pulumi/pulumi-terraform-bridge/pf/internal/schemashim"
+	pf "github.com/pulumi/pulumi-terraform-bridge/pf/tfbridge"
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge"
 )
 
-func ShimSchemaOnlyProvider(ctx context.Context, provider pfprovider.Provider) shim.Provider {
-	return &schemaOnlyProvider{
-		ctx: ctx,
-		tf:  provider,
-	}
+func shimSchemaOnlyProviderInfo(ctx context.Context, provider pf.ProviderInfo) tfbridge.ProviderInfo {
+	shimProvider := schemashim.ShimSchemaOnlyProvider(ctx, provider.NewProvider())
+	var copy tfbridge.ProviderInfo = provider.ProviderInfo
+	copy.P = shimProvider
+	return copy
 }
